+++ src/data/codec/detect/mod.rs
@@ -1,0 +1,12 @@
+// devela::data::codec::detect
+//
+//! Format detection.
+//
+
+crate::structural_mods! { // _mods
+    _mods {
+        // pub use super::{
+        //     __::_all::*,
+        // };
+    }
+}