+++ src/data/codec/encode/mod.rs
@@ -3,9 +3,10 @@
 #![doc = crate::_DOC_DATA_CODEC_ENCODE!()] // public
 #![doc = crate::_doc!(modules: crate::data::codec; encode)]
 #![doc = crate::_doc!(flat:"data")]
-#![doc = crate::_doc!(extends: encode)]
+#![doc = crate::_doc!(hr)]
 //
 
+// mod deser; // WIP Format-neutral value (de)serialization
 mod enums; // EncodingMode
 
 // WIP
@@ -19,6 +20,7 @@ mod enums; // EncodingMode
 crate::structural_mods! { // _mods
     _mods {
         pub use super::{
+            // deser::_all::*,
             enums::*,
         };
     }