+++ src/data/codec/frame/mod.rs
@@ -3,7 +3,7 @@
 #![doc = crate::_DOC_DATA_CODEC_FRAME!()] // public
 #![doc = crate::_doc!(modules: crate::data::codec; frame)]
 #![doc = crate::_doc!(flat:"data")]
-#![doc = crate::_doc!(extends: frame)]
+#![doc = crate::_doc!(hr)]
 //
 // > Where a byte unit begins and ends