+++ src/data/codec/pack/archive/mod.rs
@@ -1,0 +1,16 @@
+// devela::data::codec::pack::archive
+//
+//! Multi-resource packages.
+//
+
+// mod tar;
+// mod zip;
+
+crate::structural_mods! { // _mods
+    _mods {
+        // pub use super::{
+        //     tar::_all::*,
+        //     zip::_all::*,
+        // };
+    }
+}