+++ src/data/codec/pack/wrap/mod.rs
@@ -1,0 +1,16 @@
+// devela::data::codec::pack::wrap
+//
+//! Structured wrappers and chunked containers.
+//
+
+// mod riff;
+// mod png_chunk;
+
+crate::structural_mods! { // _mods
+    _mods {
+        // pub use super::{
+        //     riff::_all::*,
+        //     png_chunk::*,
+        // };
+    }
+}