+++ src/data/codec/symbol/mod.rs
@@ -1,0 +1,26 @@
+// devela::data::codec::symbol
+//
+//! Symbolic codes that encode data into visual marks.
+//
+
+// mod aztec;
+// mod code128;
+// mod datamatrix;
+// mod ean;
+// mod pdf417;
+// mod qr;
+// mod upc;
+
+crate::structural_mods! { // _mods
+    _mods {
+        // pub use super::{
+        //     aztec::_all::*,
+        //     code128::_all::*,
+        //     datamatrix::_all::*,
+        //     ean::_all::*,
+        //     pdf417::_all::*,
+        //     qr::_all::*,
+        //     upc::_all::*,
+        // };
+    }
+}