+++ src/data/value/mod.rs
@@ -11,8 +11,8 @@ mod of; // Oneof
 
 // mod decode; // ValueDecode WIP
 mod kind; // ValueKind, ValueKind4, WIP ValueKindSet
+// mod schema; // WIP Schemas for encoded data structures.
 // mod value; // value! WIP
-
 #[cfg(feature = "_tuple")]
 pub mod tuple; // Tuple, TupleFmt, TupleEnumRef, TupleEnumMut
 
@@ -23,6 +23,7 @@ crate::structural_mods! { // _mods, _pub_mods
             // decode::*,
             of::_all::*,
             kind::*,
+            // schema::_all::*,
             // value::*,
         };
     }