+++ src/data/value/schema/deser/mod.rs
@@ -1,0 +1,15 @@
+// devela::data::value::schema::deser
+//
+//! Format-neutral value serialization.
+//
+// > how structured values are assembled/disassembled
+
+// mod traits; // WIP
+
+crate::structural_mods! { // _mods
+    _mods {
+        // pub use super::{
+        //     traits::*,
+        // };
+    }
+}